+++ cryptoCurrency/src/main/java/org/example/cryptocurrency/service/contracts/CryptoService.java
@@ -3,6 +3,7 @@ package org.example.cryptocurrency.service.contracts;
 import org.example.cryptocurrency.models.Crypto;
 import org.example.cryptocurrency.models.UserAccount;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CryptoService {
@@ -10,6 +11,6 @@ public interface CryptoService {
     Crypto getById(Integer id);
     Crypto getBySymbol(String symbol);
     Crypto create(Crypto crypto);
-    Crypto update(String symbol, String name, double price);
+    Crypto update(String symbol, String name, BigDecimal price);
     void deleteCrypto(Integer id);
 }
