+++ cryptoCurrency/src/main/java/org/example/cryptocurrency/service/contracts/UserAccountService.java
@@ -2,11 +2,12 @@ package org.example.cryptocurrency.service.contracts;
 
 import org.example.cryptocurrency.models.UserAccount;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserAccountService {
     UserAccount findUser(Integer id);
     List<UserAccount> findAllUsers();
     UserAccount createUser(UserAccount user);
-    public UserAccount updateUserBalance(Integer userId, double newBalance);
+    public UserAccount updateUserBalance(Integer userId, BigDecimal newBalance);
 }
